Update README.md
Data is provided via a pandas dataframe.
The user supplies the page dimensions, margins, and label dimensions, all in points(1/72").
User also supplies positioning, relative to lower left corner of label, as well as width, height, font, and font size, for each field.
If design_mode is True, outline of each label and field is printed, and clipping is off. Otherwise, text is clipped, and borders do not print.